Moved ERF schemas under a new erf directory.
