+++ orchestra/google/gmp/example_dags/multiple_partners_erf_upload_bq_dag.py
@@ -22,7 +22,7 @@ from datetime import datetime
 from datetime import timedelta
 from airflow import DAG
 from airflow import models
-from orchestra.google.gmp.utils.schema import Entity_Schema_Lookup
+from orchestra.google.gmp.utils.schema.erf import Entity_Schema_Lookup
 from orchestra.google.gmp.operators.gmp_dv360_operator import DisplayVideo360MultiERFUploadBQOperator
